refactor: CheckEventTarget.js 파일 분리
- productAddMenu 관련된 부분 CheckProductAddMenu.js 로 파일 분리
